+++ src/functions/counting/combination.js
@@ -19,7 +19,7 @@ class LimeFunctionCombination extends LimeFunction {
     this.algorithms.set('r(arg{int[2]})', (step) => {
       const [l, r] = step.right.places;
       if (l.value < 0 || r.value < 0) {
-        throw new Error('error:negativePermutation');
+        throw new Error('warn:negativePermutation');
       }
       step.rus(this.lime.direct(['npr', '(', l, ',', r, ')', '/', 'npr', '(', r, ',', r, ')']));
     });